setting #32 - Extension 분리
